requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
